pyproject.toml
[project]
name = "bocco-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "emo-platform-api-sdk>=1.0.3",
    "python-dotenv>=1.2.2",
]
